Small bugfixes.
Changed a few minutes to a while after testing it on a slower computer
took a good 20 minutes to finish getting updates.

Made it so while uninstalling updates it says which update it's
uninstalling out of how many are left (it will print "Removing KB3035583
(1/6)" for example).

Also fixed how long a deleted file line can be, because my VM never
actually had any windows 10 data there so I hadn't tested it.